sync & default values
